+++ src/arraydata.h
@@ -26,7 +26,7 @@ namespace sqf
 		arraydata(std::vector<std::shared_ptr<value>> v) : mvalue(std::vector<std::shared_ptr<value>>(v)) {}
 		virtual std::string tosqf() const;
 		std::shared_ptr<value>& operator[](int index) { return mvalue.at(index); }
-		std::shared_ptr<value> operator[](int index) const { return index < 0 || index >= (int)mvalue.size() ? std::make_shared<value>() : mvalue[index]; }
+		std::shared_ptr<value> operator[](int index) const { return index < 0 || index >= static_cast<int>(mvalue.size()) ? std::make_shared<value>() : mvalue[index]; }
 		std::shared_ptr<value>& at(int index) { return mvalue.at(index); }
 		const std::shared_ptr<value> at(int index) const { return mvalue.at(index); }
 		size_t size() const { return mvalue.size(); }
@@ -38,7 +38,7 @@ namespace sqf
 
 		void push_back(std::shared_ptr<value> val) { mvalue.push_back(val); }
 		std::shared_ptr<value> pop_back() { auto back = mvalue.back(); mvalue.pop_back(); return back; }
-		void resize(int newsize);
+		void resize(size_t newsize);
 		void reverse();
 		void extend(std::vector<std::shared_ptr<value>> other);
 		void delete_at(int position);
